% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startmarvis.R
\name{startMARVIS_Web}
\alias{startMARVIS_Web}
\title{Launch MARVIS interactive User Interface in the web browser}
\usage{
startMARVIS_Web()
}
\value{
Nothing
}
\description{
startMARVIS_Web starts loads the web browser an interactive user interface built using R shiny.
}
\details{
The purpose of building the interactive user interface is to provide an easy for people who are learning how to do their first meta-analysis and/or are not comfortable with the R command line system. Includes example data for testing out a meta-analysis.
}
\examples{
\dontrun{
startMARVIS_Web()
}
}
\keyword{MARVIS}
